+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.0b53 (Unreleased)
 
 ### Features Added
+- Accept bitmask from microsoft distro and OR with azure monitor feature sdkstats
+  ([#47031](https://github.com/Azure/azure-sdk-for-python/pull/47031))
 
 ### Breaking Changes